requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
Pillow==9.0.1
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.0.0
asgiref==3.5.0
dj-database-url==0.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
Pillow==9.0.1
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.0.0
